Performance improvement
Apparently, the function islandora_handle_query_for_handle takes a long
time (about 6 seconds per object) to find out if the object has a
handle, because it checks this with the handle.net server. Now the
check is done in the MODS datastream to find if the object has a handle.