Fixed memory problem
With large collections large amounts of memory were used. This was
because all objects were loaded into memory and kept there.
Reimplemented this by only keeping the object ids in memory and loading
the object itself when it is needed. After it is used, we expose of it
in memory.